added first last lp hp and no_append options. Still only works on MRC stacks, refs #2997

+++ appion/bin/approc2d.py
@@ -2,10 +2,12 @@
 from pyami import imagefun
 from pyami import mrc
 from pyami import imagic
+from appionlib.apImage import imagefilter
 import argparse #apparently optparse is being deprecated. Barf!
 import os
 import sys
 import glob
+import copy
 
 #not inheriting Appion base classes because options are unique and stand alone
 class ApProc2d ():
@@ -22,12 +24,13 @@ class ApProc2d ():
 		
 		parser.add_argument( 'in_stack', metavar='STRING', help='Input stack')
 		parser.add_argument( 'out_stack', metavar='STRING', help='Output stack')
-		parser.add_argument( '--lp', dest='lp', type=float, metavar='FLOAT', help='Low pass filter to provided resolution. In Angstroms. ')
-		parser.add_argument( '--hp', dest='hp', type=float, metavar='FLOAT', help='High pass filter to provided resolution. In Angstroms. ')
-		parser.add_argument( '--apix', dest='apix', type=float, metavar='FLOAT', help='High pass filter to provided resolution. In Angstroms. ')
+		parser.add_argument( '--lp', dest='lp', default=None, type=float, metavar='FLOAT', help='Low pass filter to provided resolution. In Angstroms. ')
+		parser.add_argument( '--hp', dest='hp', default=None, type=float, metavar='FLOAT', help='High pass filter to provided resolution. In Angstroms. ')
+		parser.add_argument( '--apix', dest='apix', default=1.0, type=float, metavar='FLOAT', help='High pass filter to provided resolution. In Angstroms. ')
 		parser.add_argument("--invert", dest="invert", help="Invert contrast", action='store_true', default=False )
-		parser.add_argument('--first', dest='first', type=int,metavar='FLOAT', help="Operate on particles starting from the given INT. Note: particle numbers in stacks start with 0")
-		parser.add_argument('--last', dest='first', type=int, metavar='FLOAT', help="Operate on particles starting from the given INT. Note: particle numbers in stacks start with 0")
+		parser.add_argument('--first', dest='first', default=0, type=int,metavar='FLOAT', help="Operate on particles starting from the given INT. Note: particle numbers in stacks start with 0")
+		parser.add_argument('--last', dest='last', default=0, type=int, metavar='FLOAT', help="Operate on particles starting from the given INT. Note: particle numbers in stacks start with 0")
+		parser.add_argument('--no_append', dest='no_append', default=False, action='store_true', help="Overwrite preexisting output stack")
 		self.args=parser.parse_args()
 		#print self.args
 	
@@ -37,25 +40,51 @@ class ApProc2d ():
 		print instacktype
 		print "Parsing header"
 		if instacktype=='.mrc':
-			header=mrc.readHeaderFromFile(self.args.in_stack)
+			inheader=mrc.readHeaderFromFile(self.args.in_stack)
 		elif instacktype=='.hed' or instacktype=='.img':
-			header=imagic.readImagicHeader(self.args.in_stack)
+			inheader=imagic.readImagicHeader(self.args.in_stack)
 			
-		print header
-		nptcls=header['nz']
-		print "Nptcls = ", header['nz']
+		print inheader
+		in_nptcls=inheader['nz']
+		print "Nptcls = ", inheader['nz']
 		
 		### First steps
-		fdat=open(self.args.in_stack,'r')
-		fout=open(self.args.out_stack,'wb')
-		headerbytes=mrc.makeHeaderData(header)
-		fout.write(headerbytes)
 		
-		for n in range(nptcls):
-			ptcl=mrc.readDataFromFile(fdat,header,n)
+		# determine nptcls to add
+		if self.args.last is not None:
+			add_nptcls=self.args.last-self.args.first+1
+		else:
+			add_nptcls=in_nptcls-self.args.first
+
+		fdat=open(self.args.in_stack,'rb')
+		
+		if os.path.exists(self.args.out_stack) and self.args.no_append is False:
+			outheader=mrc.readHeaderFromFile(self.args.out_stack)
+			out_nptcls=outheader['nz']
+			outheader['nz']=out_nptcls+add_nptcls #create header with correct number of particles after processing
+			fout=open(self.args.out_stack,'rb+')
+		else:
+			fout=open(self.args.out_stack,'wb')
+			outheader=copy.deepcopy(inheader)
+			outheader['nz']=add_nptcls
+			out_nptcls=0
+		outheaderbytes=mrc.makeHeaderData(outheader)
+
+		fout.write(outheaderbytes)
+		
+			
+		for n in range(self.args.first, self.args.first+add_nptcls):
+			ptcl=mrc.readDataFromFile(fdat,inheader,n)
 			if self.args.invert is True:
 				ptcl=ptcl*-1
+			if self.args.lp is not None:
+				ptcl=imagefilter.lowPassFilter(ptcl,apix=self.args.apix, radius=self.args.lp)
+			if self.args.hp is not None:
+				ptcl=imagefilter.highPassFilter2(ptcl, self.args.hp, apix=self.args.apix)
 			mrc.appendArray(ptcl,fout)
+			out_nptcls+=1
+		
+		print "Total particles = ", out_nptcls
 		fout.close()
 		
 		### Last steps
